alpha-stack/requirements.txt
websocket-client==1.8.0
requests==2.32.3
flask==3.0.3
